Update MiscConversionTests.cs
Add new test case for 43kW AC charging, should flag as level 3, since it's >40kW.
This makes it consistent with imported data.

Yes. I know Level 3 is not really defined by SAE for either DC or AC, but we've been using it to denote charging at power levels of 40kW and above.